pyproject.toml
[tool.poetry]
name = "machine-learning-source-code-plagiarism"
version = "0.1.0"
description = ""
authors = ["Bosire Nyakundi <nyakundi@hotmail.co.uk>"]
readme = "README.md"
packages = [{include = "machine_learning_source_code_plagiarism"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
six = "^1.16.0"
javalang = "^0.13.0"
pandas = "^2.1.1"
lizard = "^1.17.10"
pyspellchecker = "^0.8.1"
scikit-learn = "^1.4.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
